Employ *latency* compensation mechanism. :)
This commit adds a stable latency calculation which accounts for:
-default playback device latency
-pyaudio buffers

and ensures that this receiver falls into synchronization with other
playing devices.

We actually *want* the play-head to be offset and ahead. This way, it
synchronizes with other Airplay players (this is why Airplay clients
stream a bunch of RTP in advance). In my tests with other Sonos
endpoints - this latency calculation puts everything in sync to within
a millisecond.

Admittedly, the output is not phase locked, but PTP may help to take
care of that.